Fix BlockTokenIdentifier password computation between Namenode HDP3 and datanodes HDP2
When a datanode registers the the namenode, the namenode transmits a set of keys, that consists in an id and a random long value. Those keys are used as seed to a cryptographic function to compute hashes.

When the namenode generates a BlockTokenIdentifier it does the following:
  - serialize it with its write method
  - get one the keys
  - compute a hash with this key as seed
  - send the BlockTokenIdentifier as binary stream to the datanode, along with the key id used and the expected hash

The datanode receiving the BlockTokenIdentifier does the following:
  - deserialize the BlockTokenIdentifer
  - get the key id
  - reserialize the BlockTokenIdentifier
  - compute the hash with its own copy of the key (in memory)
  - verify the computed hash against the exepected one

With HDP3, the namenode sends additional data in the byte array. The HDP2 datanode ignores this data, so when it reserializes the BlockTokenIdentifier, the byte array does not correspond to the one the namenode used to compute the hash. Thus, even if the datanode has the correct keys, it is not able to get the same hash.

The fix consist in deserializing/serializing this additional data, if it exists. Hashes are then correctly computed, wether the namenode is HDP3 or HDP2.