Fix HeartbeatResponse between Namenode HDP 3 and datanodes HDP 2 (#137)
HeartbeatResponseProto messages have different layouts between HDP3 and HDP2, they use the same fields but some indexes are switched. That leads HDP2 datanode to incorrectly parse the fullBlockReportLeaseId, and as a consequence, datanodes cannot perform full block reports.

The workaround relies on the fact that protobuf is able to detect the presence of unknown fields in a message. Thus when a datanode HDP 2 receives a HeartbeatResponse from a Namenode HDP 3, we can detect the presence of such unknown field. In this case, we can reserialize the message with the correct protobuf message and access all expected fields.

This mecahnism is implemented through a wrapper class, as protobuf messages uses final class, polymorphism was not an option.

Co-authored-by: William Montaz <w.montaz@criteo.com>